Implement support for config includes/imports
This adds support for "importing" or "including" configuration
files by specifying a config key which, when encountered, is to be
treated as a list of config files to include and be merged into the
source config file, creating a config hierarchy.

This allows developers to have e.g., a "base" configuration which
is then imported by their "beta" config or "production" config to
be overridden or extended.

One can call `SetConfigIncludeKey("includes")` and then
include config files in their e.g., YAML config as follows:

```yaml
includes:
  - "base"
```